Refactoring code, fix work with endpoints.

+++ src/main/java/com/example/internship/facade/BookcaseFacade.java
@@ -20,7 +20,7 @@ public class BookcaseFacade {
     }
 
     public void delete(long id) {
-        bookcaseService.delete((int) id);
+        bookcaseService.delete(id);
     }
 
     public BookcaseCreateResponse savePostRequest(BookcaseCreateRequest bookcaseCreateRequest){
